pyproject.toml
[project]
name = "my-silicon-brain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.69.0",
    "arxiv>=2.2.0",
    "mcp>=1.17.0",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.1.1",
    "pymilvus>=2.4.0",
    "openai>=1.0.0",
    "unstructured>=0.11.0",
    "watchdog>=3.0.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.1.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "feedparser>=6.0.0"
]
